Use BehaviourSubject and listen to profile stream
The UserService has app scope and a BehaviourSubject seeded with an
empty map. The UserService exposes the BehaviourSubject stream, listens
to firestore's snapshots stream and adds events to the BehaviourSubject.

The firestore's snapshots stream only emits once then on any updates
so we can now listen to the UserService's BehaviourSubject's stream
at any time and get latest profile data without performing unnecessary
reads.

The AuthService now sets up the UserService listening to firestore's
snapshots when the user signs in or opens the app already signed in.
Snapshots are then added to the UserService's BehaviourSubject.