Edit Profile (#17)
* Add merge option to firestore setDoc

This allows an update that creates the doc if none exists

* Use image_picker's maxHeight to resize image

Add a UserService that holds the profileDocsStream, made up of a Stream
from the FirestoreService and the uid from the AuthService.

Update db to have profile docs rather than profilePics.

* Add ImagesService & refactor

* Add size and loading state to Avatar

* Use BehaviourSubject and listen to profile stream

The UserService has app scope and a BehaviourSubject seeded with an
empty map. The UserService exposes the BehaviourSubject stream, listens
to firestore's snapshots stream and adds events to the BehaviourSubject.

The firestore's snapshots stream only emits once then on any updates
so we can now listen to the UserService's BehaviourSubject's stream
at any time and get latest profile data without performing unnecessary
reads.

The AuthService now sets up the UserService listening to firestore's
snapshots when the user signs in or opens the app already signed in.
Snapshots are then added to the UserService's BehaviourSubject.

* Profile name updating

* Fix edit profile screen overflowing from keyboard

I put the edit profile screen in a SingleChildScrollView

* Update Screen: ImagePicker to EditProfile

EditProfileScreen better captures the role of the widget.