Sign out
Calling dispose on the UserService in the AuthService's signOut cancels
the subscription to the firestore stream otherwise we get an error about
trying to listen to the firestore without being authenticated. This
works but it might be good to merge AuthService into UserService to
avoid this complexity and tight coupling.