Create a VenueIcon widget, save as image, upload
I used screenshot package to get png bytes from a
widget.

We hide icon widget behind the main photo in a
stack.

We screenshot the icon widget then upload as bytes
rather than saving to file, then retrieve the bytes
and create a Marker with a BitmapDescriptor from
the bytes and add the Marker to the map.