Use a stack with a marker icon in Add Venue
This is to avoid the latency that comes with calling setState
on every cameraMoved callback.

I've used a LayoutBuilder with maxHeight and maxWidth (not divided by
2) and a Positioned moving the marker up 40. It's a weird solution -
I've  left the old marker drawn by setstate in there for now to check if
the two line up in different phones with different dimensions.