On team request add pending status (#86)
A cloud function triggers on a user adding a doc with a team mate's
id to their team-requests collection, then adds the user's id to the
potential team mate's "pendingTeamRequests" list.

The "pendingTeamRequests" list is used by the PlayerProfileScreen
to check if the "Team Up" button should be disabled and set to
"Pending...".

Part of #81