Edit Profile in user's Profile Screen (#120)
Tapping your own profile pic to go to your Profile Screen. The
owner of the profile gets a different UI, with edit icons on
the profile pic and name that let you edit the values.

The edit name and edit profile pic screens are now re-used in
onboarding and editing from the profile.

Closes #116