Update Notifications badge (#136)
* Update notifications badge on app resume

On app resume, get an aggregated COUNT query of unviewed
Notifications and update the Icon badge. We keep a stream controller
and have the badge part of the UI subscribe to the stream, then
read the COUNT and add to the stream.

As each Notification is built, if viewed is false, update the DB with
true.

* Update version & ios min deployment target

* Update cloud functions to edit notification with viewed false

* React to foregound notifications

If NotificationsScreen is not open, add a badge to the
NavigationDestination with the number of unviewed notifications

Part of #82