Keep tokens fresh (#147)
* Check for fresh tokens and update

To ensure the client always updates the relevant firestore doc with the
latest token ASAP, we use shared preferences to keep the most recent
token and check against the token from the SDK:
- In `onTokenRefresh`
- on sign in (in case `onTokenRefresh` is called while not signed in)
- in `init()` at the end of onboarding

* Add cloud function that updates tokens everywhere

For now, the only place there are follower tokens is in profiles but
we will probably have them in other places too. We can update the
cloud function as needed.

* Catch error on sending fcm and print

We don't remove old tokens as we rely on them to find profiles that need
updating when a token refresh occurs on the client and we update the
fcmTokens/{playerId} doc.