Create messaging screen
Made a screen with a StreamBuilder and ListView for messages.

Added a 'Message' button to PlayerProfile that finds or creates the
unique conversation between two players and opens the MessagesScreen.

Added a cloud function that takes a message and a conversation id,
adds the message to the conversation and notifies everyone.

Switched to using player ids rather than tokens (less complexity,
we can look at reducing cost later).