Complete ConversationsScreen
Move image and player caches out into main and inject into repositories
layer, so we have just the two caches.

Completed ConversationsScreen by creating a Conversation model and a
ConversationViewModel that makes db and storage retrievals to get all
the data needed for the view (in this case the ConversationWidget). The
function retrieveConversationViewModels in the ConversationsService does
all the work then returns the list of ConversationViewModels. I was
exploring this type of async data retrieval to avoid the problem of the
listview.builder jumping around when the FutureBuilder inside the items
returns its value. It means waiting for everything before we show
anything which is going to create a terrible UX.

A conversation between 2 players has the id of the 2 players combined in
alphabetic order so can always be calculated, in case either player taps
'Message' from the ProfileScreen.

I made the ConversationWidget show the name and profile pic of the other
person along with a truncated value of the last message that was sent.