Use ViewModels and pre-cache
Our models represent what is in the database and the view models
contain everything needed by the view, which may require further async
calls.

Each Notification widget now takes a ViewModel object and all async work
is performed before the ListView widget is built. This avoids problems
with the list size changing with async builds but creates a long wait.

Added caches for images and player profiles that cache for a given
amount of time (default 15 secs).

Used the cloud_firestore Timestamp in Notification objects. I would have
preferred to avoid the dependency but the json already holds a Timestamp
object