Use ViewModels and pre-cache (#170)
* Remove return in sendMessageToParticipants

The callable cloud function sendMessageToParticipants is having
errors in the logs (and resulting in Crashlytics errors) but all logs
are being printed so we have removed the return statement to see
if that was the problem.

* Increase build number and version

Onve a version is on test devices we should always increase the build
number and version so we can differentiate crashes while debugging from
crashes on devices in release builds.

* Create NotificationsService

In order to hold everything relevant to Notifications. The naming is
unfortunate but we need to remember that the MessagingService holds
member variables and functions for notifications (small 'n' for
FCM notifications).

* Use ViewModels and pre-cache

Our models represent what is in the database and the view models
contain everything needed by the view, which may require further async
calls.

Each Notification widget now takes a ViewModel object and all async work
is performed before the ListView widget is built. This avoids problems
with the list size changing with async builds but creates a long wait.

Added caches for images and player profiles that cache for a given
amount of time (default 15 secs).

Used the cloud_firestore Timestamp in Notification objects. I would have
preferred to avoid the dependency but the json already holds a Timestamp
object