Access the cache iteratively (#172)
* Access the cache iteratively

Using Future.wait seems to have been the problem - perhaps it puts
each callback in the event queue immediately, meaning each cache access
is a miss, creating a long delay.

Counterintuitively, by awaiting each cache access one at a time we get
much better latency as each each cache access will be a cache hit on the
second access.

* Remove return from callable callable function

`sendMessageToParticipants` was returning an object with the results of
sending FCM messages.

I'm not sure if we just shouldn't return anything or if there was a bug
in the code but we don't use the returned value anyway.

* Upgrade flutter and ios dependencies