Resize images in a cloud function (#180)
* Add a resize-images function

* Setup jest and get a test running

* Add files to tsconfig's "include" list

The relevant files were giving a parsing error as apparently
ESLint was configured to run on them using `parserOptions.project`.

Finish setting up jest

* Fix where index.js is found

The build is now building to lib/src/ and lib/test so we needed to set
the 'main' entry of the package.json to look in 'src'

Fix resize-image.test.ts

In order for the emulated cloud function to run, it appears that the
bucket needs to be named the same as the live bucket.

* Debug the firebase functions

I added a launch configuration that attaches to the emulators process.

The emulators must be started with the `--inspect-functions` flag, and I
added notes to the README.md on how to do so.

* Update node and npm packages for functions

* Remove image resizing on the client

Instead, we let the cloud function do the resizing.

We now have small, medium and large images created on each image upload.

We use the .jpg extension so the content type is correctly detected.

Closes #179