Keep track of uploaded images using a timestamp (#188)
* Use NetworkImage for player profile pic

Update PlayersService's `retrieveProfilePic` to create the relevant Url
string for a NetworkImage rather than downloading the bytes with the
Storage object.

Keep constants for database and bucket names in a globals file.

* Change Conversations list items to use AsyncAvatar

* Use AsyncAvatar in Notifications widgets

* Remove ImageBytesCache

We now rely on caching from Cloud Storge and Flutter's NetworkImage.

* Save and retrieve profile pic with a timestamp

Update AsyncAvatar to retrieve the Player model then use the
picTimestamp to retrieve the profile pic.

Save profile pics using a timestamp as the filename.

* Downgrade eslint to fix errors

eslint after 9.0 requires migrating and the file changes threw errors
when I deployed that I couldn't figure out a fix for, seemed easier just
to downgrade for now.

* Use file path to update db in resize-images

We upload images to (eg.) profiles/playerId/timestamp.jpg then in the
cloud function we separate into 'profiles', playerId & timestamp.

We use the separated values to update the database so the profile/venue
can track the images that have been uploaded and the status of resizing.

* Bust the player cache in PlayerProfileScreen

We bust the cache when a new Player is emitted so we don't
get a stale Player that doesn't have changes like `picStatus`.

* Upgrade packages and pods

* Have 2 separate resize-images functions

The resize-images functions run on our different buckets - a
bucket in the us region and one in the aus region.

We also write to the appropriate firestore database, ie. we use the
firestore in the same region as the bucket and function.